reorganized project structure
