+++ jobs/Backend/Task/DTO/Currency.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.DTO
 {
     public class Currency
     {
