+++ jobs/Backend/Task/DTO/ExchangeRate.cs
@@ -1,4 +1,4 @@
-﻿namespace ExchangeRateUpdater
+﻿namespace ExchangeRateUpdater.DTO
 {
     public class ExchangeRate
     {
