+++ jobs/Backend/Task/DataSource/IExchangeRateDataSourceProvider.cs
@@ -1,0 +1,7 @@
+﻿namespace ExchangeRateUpdater.DataSource
+{
+    public interface IExchangeRateDataSourceProvider
+    {
+        string Get();
+    }
+}
