+++ jobs/Backend/Task/Program.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using ExchangeRateUpdater.DTO;
+using ExchangeRateUpdater.Provider;
 using log4net;
 using log4net.Config;
 using Microsoft.Extensions.Configuration;
